draft 1, Laura
I have started the html for the brand page and added the css and bootstrap.
Q for Raphael: how do we make sure the css will be the same on all of our htmls, if the css will later be changed? Because you work on your html offline, and there you add the css styles and such. Will it automatically understand that the css file you are refering to in your html is the one that is here on github?